input form data for edit and stats component
