--- src/components/Admin/Trainings/Stats/WeekStats.css
@@ -1,3 +1,0 @@
-.weeklyStatSelectorContainer {
-    float: right;
-}
